modules/local/idmapping/requirements.txt
gprofiler-official==1.0.0
